requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymupdf==1.23.7
python-docx==1.1.0
python-pptx==0.6.23
piper-tts==1.2.0
